+++ tapo/src/api/child_devices/plug_power_strip_handler.rs
@@ -1,16 +1,15 @@
 use crate::api::PowerStripHandler;
 use crate::error::{Error, TapoResponseError};
 use crate::requests::{EmptyParams, GenericSetDeviceInfoParams, TapoParams, TapoRequest};
-use crate::responses::{DecodableResultExt, P300ChildResult};
+use crate::responses::{DecodableResultExt, PlugPowerStripResult};
 
-
-/// Handler for the [P300](https://www.tapo.com/en/search/?q=T100) child devices.
-pub struct P300ChildHandler<'h> {
+/// Handler for the [P300](https://www.tapo.com/en/search/?q=P300) child plugs.
+pub struct PlugPowerStripHandler<'h> {
     power_strip_handler: &'h PowerStripHandler,
     device_id: String,
 }
 
-impl<'h> P300ChildHandler<'h> {
+impl<'h> PlugPowerStripHandler<'h> {
     pub(crate) fn new(power_strip_handler: &'h PowerStripHandler, device_id: String) -> Self {
         Self {
             power_strip_handler,
@@ -18,13 +17,13 @@ impl<'h> P300ChildHandler<'h> {
         }
     }
 
-    /// Returns *device info* as [`P300ChildResult`].
+    /// Returns *device info* as [`PlugPowerStripResult`].
     /// It is not guaranteed to contain all the properties returned from the Tapo API.
-    pub async fn get_device_info(&self) -> Result<P300ChildResult, Error> {
+    pub async fn get_device_info(&self) -> Result<PlugPowerStripResult, Error> {
         let request = TapoRequest::GetDeviceInfo(TapoParams::new(EmptyParams));
 
         self.power_strip_handler
-            .control_child::<P300ChildResult>(self.device_id.clone(), request)
+            .control_child::<PlugPowerStripResult>(self.device_id.clone(), request)
             .await?
             .ok_or_else(|| Error::Tapo(TapoResponseError::EmptyResult))
             .map(|result| result.decode())?
@@ -46,7 +45,6 @@ impl<'h> P300ChildHandler<'h> {
         let json = serde_json::to_value(GenericSetDeviceInfoParams::device_on(true)?)?;
         let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
 
-
         self.power_strip_handler
             .control_child::<serde_json::Value>(self.device_id.clone(), request)
             .await?;
@@ -58,7 +56,6 @@ impl<'h> P300ChildHandler<'h> {
     pub async fn off(&self) -> Result<(), Error> {
         let json = serde_json::to_value(GenericSetDeviceInfoParams::device_on(false)?)?;
         let request = TapoRequest::SetDeviceInfo(Box::new(TapoParams::new(json)));
-
 
         self.power_strip_handler
             .control_child::<serde_json::Value>(self.device_id.clone(), request)