+++ tapo/src/responses/child_device_list_hub_result.rs
@@ -4,7 +4,6 @@ mod t100_result;
 mod t110_result;
 mod t300_result;
 mod t31x_result;
-mod p300_child_result;
 
 pub use ke100_result::*;
 pub use s200b_result::*;
@@ -12,24 +11,23 @@ pub use t100_result::*;
 pub use t110_result::*;
 pub use t300_result::*;
 pub use t31x_result::*;
-pub use p300_child_result::*;
 
 use serde::{Deserialize, Serialize};
 
 use crate::error::Error;
 use crate::responses::{DecodableResultExt, TapoResponseExt};
 
-/// Child device list result.
+/// Hub child device list result.
 #[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ChildDeviceListResult {
-    /// Child devices
+pub(crate) struct ChildDeviceListHubResult {
+    /// Hub child devices
     #[serde(rename = "child_device_list")]
-    pub devices: Vec<ChildDeviceResult>,
+    pub devices: Vec<ChildDeviceHubResult>,
 }
 
-impl DecodableResultExt for ChildDeviceListResult {
+impl DecodableResultExt for ChildDeviceListHubResult {
     fn decode(self) -> Result<Self, Error> {
-        Ok(ChildDeviceListResult {
+        Ok(ChildDeviceListHubResult {
             devices: self
                 .devices
                 .into_iter()
@@ -39,7 +37,7 @@ impl DecodableResultExt for ChildDeviceListResult {
     }
 }
 
-impl TapoResponseExt for ChildDeviceListResult {}
+impl TapoResponseExt for ChildDeviceListHubResult {}
 
 /// Device status.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -50,10 +48,10 @@ pub enum Status {
     Offline,
 }
 
-/// Child device result.
+/// Hub child device result.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(tag = "model")]
-pub enum ChildDeviceResult {
+pub enum ChildDeviceHubResult {
     /// KE100 thermostatic radiator valve (TRV).
     KE100(Box<KE100Result>),
     /// S200B button switch.
@@ -68,42 +66,37 @@ pub enum ChildDeviceResult {
     T310(Box<T31XResult>),
     /// T315 temperature & humidity sensor.
     T315(Box<T31XResult>),
-    // P300 child plug.
-    P300(Box<P300ChildResult>),
     /// Catch-all for currently unsupported devices.
     /// Please open an issue if you need support for a new device.
     #[serde(other)]
     Other,
 }
 
-impl DecodableResultExt for ChildDeviceResult {
+impl DecodableResultExt for ChildDeviceHubResult {
     fn decode(self) -> Result<Self, Error> {
         match self {
-            ChildDeviceResult::KE100(device) => {
-                Ok(ChildDeviceResult::KE100(Box::new(device.decode()?)))
+            ChildDeviceHubResult::KE100(device) => {
+                Ok(ChildDeviceHubResult::KE100(Box::new(device.decode()?)))
             }
-            ChildDeviceResult::S200B(device) => {
-                Ok(ChildDeviceResult::S200B(Box::new(device.decode()?)))
+            ChildDeviceHubResult::S200B(device) => {
+                Ok(ChildDeviceHubResult::S200B(Box::new(device.decode()?)))
             }
-            ChildDeviceResult::T100(device) => {
-                Ok(ChildDeviceResult::T100(Box::new(device.decode()?)))
+            ChildDeviceHubResult::T100(device) => {
+                Ok(ChildDeviceHubResult::T100(Box::new(device.decode()?)))
             }
-            ChildDeviceResult::T110(device) => {
-                Ok(ChildDeviceResult::T110(Box::new(device.decode()?)))
+            ChildDeviceHubResult::T110(device) => {
+                Ok(ChildDeviceHubResult::T110(Box::new(device.decode()?)))
             }
-            ChildDeviceResult::T300(device) => {
-                Ok(ChildDeviceResult::T300(Box::new(device.decode()?)))
+            ChildDeviceHubResult::T300(device) => {
+                Ok(ChildDeviceHubResult::T300(Box::new(device.decode()?)))
             }
-            ChildDeviceResult::T310(device) => {
-                Ok(ChildDeviceResult::T310(Box::new(device.decode()?)))
+            ChildDeviceHubResult::T310(device) => {
+                Ok(ChildDeviceHubResult::T310(Box::new(device.decode()?)))
             }
-            ChildDeviceResult::T315(device) => {
-                Ok(ChildDeviceResult::T315(Box::new(device.decode()?)))
+            ChildDeviceHubResult::T315(device) => {
+                Ok(ChildDeviceHubResult::T315(Box::new(device.decode()?)))
             }
-            ChildDeviceResult::P300(device) => {
-                Ok(ChildDeviceResult::P300(Box::new(device.decode()?)))
-            }
-            ChildDeviceResult::Other => Ok(ChildDeviceResult::Other),
+            ChildDeviceHubResult::Other => Ok(ChildDeviceHubResult::Other),
         }
     }
 }